Refactor PromptNode, separate PromptModelInvocationLayers in providers.py
