+++ haystack/nodes/prompt/__init__.py
@@ -1,1 +1,2 @@
 from haystack.nodes.prompt.prompt_node import PromptNode, PromptTemplate, PromptModel
+from haystack.nodes.prompt.providers import PromptModelInvocationLayer
